Revert "add footer tray for mobile view (#5785)"
This reverts commit c7c0225dc0bcb37fed0273c6aa24e3fa96085655.
