+++ frontend/app/src/components/home/LeftPanel.svelte
@@ -3,7 +3,6 @@
     import { rtlStore } from "../../stores/rtl";
     import { currentTheme } from "../../theme/themes";
     import { layoutStore } from "../../stores/layout";
-    import BottomNav from "./nav/BottomNav.svelte";
 </script>
 
 <section
@@ -29,10 +28,6 @@
             on:askToSpeak
             on:hangup
             on:newChannel />
-
-        {#if $layoutStore.showBottomNav}
-            <BottomNav on:profile />
-        {/if}
     </div>
 </section>
