+++ frontend/app/src/components/home/MiddlePanel.svelte
@@ -92,7 +92,7 @@
     {#if $pathParams.kind === "explore_groups_route"}
         <RecommendedGroups {joining} on:joinGroup on:leaveGroup on:upgrade />
     {:else if $pathParams.kind === "communities_route"}
-        <ExploreCommunities on:profile on:upgrade on:createCommunity />
+        <ExploreCommunities on:upgrade on:createCommunity />
     {:else if $pathParams.kind === "admin_route"}
         {#await import("./admin/Admin.svelte")}
             <div class="loading">
