+++ frontend/app/src/components/home/nav/LeftNavItem.svelte
@@ -12,29 +12,20 @@
     export let unread = emptyUnreadCounts();
     export let video = { muted: 0, unmuted: 0 };
     export let disabled = false;
-    export let orientation: "vertical" | "horizontal" = "vertical";
 </script>
 
-<div
-    role="button"
-    tabindex="0"
-    class:separator
-    class:selected
-    class:horizontal={orientation === "horizontal"}
-    class="nav-item"
-    on:click>
+<div role="button" tabindex="0" class:separator class:selected class="left-nav-item" on:click>
     <div class="icon" title={$_(label.key, label.params)}>
         <slot />
         <UnreadCount {unread} />
         <VideoCallIcon {video} />
     </div>
-    {#if orientation === "vertical"}
-        <div class="label"><Translatable resourceKey={label} /></div>
-    {/if}
+    <div class="label"><Translatable resourceKey={label} /></div>
+    <div class="menu"><slot name="menu" /></div>
 </div>
 
 <style lang="scss">
-    :global(.nav-item .unread-count) {
+    :global(.left-nav-item .unread-count) {
         right: toRem(-9);
         top: 85%;
         @include mobile() {
@@ -45,7 +36,7 @@
     $size: toRem(48);
     $mobile-size: toRem(40);
 
-    .nav-item {
+    .left-nav-item {
         display: flex;
         align-items: center;
         gap: toRem(16);
@@ -70,15 +61,6 @@
             border-bottom: 1px solid var(--bd);
         }
 
-        &.horizontal {
-            flex-direction: column;
-            gap: $sp2;
-
-            &.selected {
-                background-color: transparent;
-            }
-        }
-
         .icon {
             flex: 0 0 $size;
             width: $size;
@@ -98,11 +80,6 @@
         .label {
             flex: auto;
             white-space: nowrap;
-        }
-
-        &.vertical .label {
-            text-transform: uppercase;
-            @include font(light, normal, fs-60);
         }
 
         .menu {
